lab1 update

+++ Lab1/src/main.cpp
@@ -19,27 +19,27 @@ int main () {
         return 0;
     }
 
-    input_value = atoi(input_string.c_str());
+    input_value = atol(input_string.c_str());
 
 
-    if (fmod (sqrt(8 * input_value + 1), 1)) {
+    if (fmodl (sqrtl(8 * input_value + 1), 1)) {
         std::cout << 0;
 	return 0;
     }
 
-    double des = 1 - 8 * (-input_value);
+    long double des = 1 - 8 * (-input_value);
 
-    double value_num = 0;
+    long double value_num = 0;
 
     if (des >= 0) { 
-        value_num = (-1 + sqrt(des)) / 2;
+        value_num = (-1 + sqrtl(des)) / 2;
     }
     else {
 	std::cout << 0;
 	return 0;
     }
 
-    if (fmod (value_num, 1)) {
+    if (fmodl (value_num, 1)) {
 	std::cout << 0;
 	return 0;
     }
